Potree cropping support
